Upgrade tonic and prost dependencies
`prost` 0.8.0 contains a security fix recommended by the maintainers[1].
The security flaw doesn't impact this crate directly as it's related
to a panic in converting to SystemTime from protobuf timestamps from
unknown sources; ya-gcp doesn't do such conversions. The upgrade
is nevertheless useful as downstream consumers may start using 0.8.0 and
this crate re-exports a few prost types, so interop will be improved.

Tonic has been upgraded to a version that uses prost 0.8.0, and as a
consequence the generated code has been updated using a build with this
new version.

[1] https://github.com/tokio-rs/prost/releases/tag/v0.8.0