Retry all Unknown statuses in PubSub
Checking the error message for an implied GOAWAY proved to be too
brittle to correctly handle that error in practice. Instead, any unknown
failure will be retried. This might result in some incorrect retries,
but that seems to be much less frequent than real triggered Unknowns
which need to be retried.