Revert pubsub codegen additions
As part of the bigtable implementation, the protobuf generated pubsub
API was updated as well. However, because prost's codegen doesn't follow
the same semver semantics as protobuf does, those were breaking changes
to the pubsub API. To help adopt other patches (#23), this change
reverts those breakages so that semver does not need a major bump.

Future changes may mitigate this by adding `#[non_exhaustive]`
attributes to the generated code. However that will be a major bump as
well, which is better left to later diffs. Another improvement would be
to more readily support breaking changes downstream, but that requires
some API work that isn't easy at this time.