Allow . in bucket names (#27)
Bucket names in GCS are allowed to contain '.' (under some conditions; see here), but tame-gcs fails validation for those names, as reported here.

This PR works around the issue by copying tame-gcs's bucket verification code, and modifying it to allow '.'. It's written with the assumption that tame-gcs will eventually fix the issue and so we can remove this code; if we intended to actually keep this around, it would probably make more sense to integrate BucketNameError into InvalidNameError.

Also, while testing this I noticed that build fails with --features=storage, because of a missing tower dependency.